requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
# Supprime psycopg2-binary et celery pour l'instant